requirements.txt
astroid==2.11.7
dill==0.3.5.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pylint==2.14.5
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
